get questions shown on Last Year tab
